+++ src/timeline.js
@@ -70,7 +70,7 @@ const EMPTY_GROUP_KEY = 'empty-group';
 // This was added by bogdan. From my understanding it reprezents the table vertical scrollbar width
 // If we don't take in consideration this, a horizontal scrollbar appears
 export const TABLE_OFFSET = 15;
-export const DEFAULT_ITEM_HEIGHT = 40;
+export const DEFAULT_ITEM_HEIGHT = 30;
 export const DEFAULT_ROW_CLASS = 'rct9k-row';
 export const DEFAULT_ROW_EVEN_CLASS = 'rct9k-row-even';
 /**
@@ -85,6 +85,8 @@ const FADE_OPACITY_OFFSET = 0.1;
 const FADE_TIMER_INTERVAR = 100;
 export const ZOOM_PERCENT = 0.2;
 export const MIN_DISPLAY_TIME = 60000;
+const DEFAULT_VERTICAL_GAP_BETWEEN_OVERLAPPING_ITEMS = 1;
+const DEFAULT_ROW_TOP_BOTTOM_PADDING = 1;
 
 export const PARENT_ELEMENT = componentId => document.querySelector(`.rct9k-id-${componentId} .parent-div`);
 
@@ -506,6 +508,22 @@ export default class Timeline extends React.Component {
     displayItemOnSeparateRowIfOverlap: PropTypes.oneOfType([PropTypes.bool, PropTypes.func]),
 
     /**
+     * The vertical gap between overlapping items displayed on separate subrows in case `displayItemOnSeparateRowIfOverlap == true`
+     *
+     * @default 1
+     * @type { number }
+     */
+    verticalGapBetweenOverlappingItems: PropTypes.number,
+
+    /**
+     * The top and down empty space inside a row
+     *
+     * @default 1
+     * @type { number }
+     */
+    rowTopBottomPadding: PropTypes.number,
+
+    /**
      * The segments with bigger index are staying in front of the ones with smaller index.
      * To not show the segment in front of the now marker, cursor time
      * please do not return a number greater than 998 @see rct9k-marker style
@@ -576,6 +594,8 @@ export default class Timeline extends React.Component {
     onSplitChange: undefined,
     splitPaneProps: undefined,
     displayItemOnSeparateRowIfOverlap: true,
+    verticalGapBetweenOverlappingItems: DEFAULT_VERTICAL_GAP_BETWEEN_OVERLAPPING_ITEMS,
+    rowTopBottomPadding: DEFAULT_ROW_TOP_BOTTOM_PADDING,
     zIndexFunction() {
       return 3;
     },
@@ -1134,7 +1154,7 @@ export default class Timeline extends React.Component {
     let that = this;
     this.rowIdToRowIndexMap = {};
     _.forEach(groups, (group, index) => {
-      totalItemsHeight += (that.rowHeightCache[group.id] || 1) * that.props.itemHeight;
+      totalItemsHeight += that.getNonEmptyRowHeight(group.id);
       this.rowIdToRowIndexMap[group.id] = index;
     });
     let heightToFillIn = this._grid.props.height - totalItemsHeight;
@@ -1868,7 +1888,7 @@ export default class Timeline extends React.Component {
       const layersInRow = this.props.rowLayers.filter(r => r.rowNumber === rowIndex);
       let rowHeight = this.props.itemHeight;
       if (this.rowHeightCache[rowId]) {
-        rowHeight = rowHeight * this.rowHeightCache[rowId];
+        rowHeight = this.getNonEmptyRowHeight(rowId);
       }
       var props = this.props;
       return (
@@ -1913,6 +1933,8 @@ export default class Timeline extends React.Component {
             this.getEndFromItem,
             timelineTestids,
             this.props.displayItemOnSeparateRowIfOverlap,
+            this.props.verticalGapBetweenOverlappingItems,
+            this.props.rowTopBottomPadding,
             this.props.zIndexFunction,
             rowIndex
           )}
@@ -1946,8 +1968,16 @@ export default class Timeline extends React.Component {
     if (group.rowHeight && group.key.startsWith(EMPTY_GROUP_KEY)) {
       return group.rowHeight;
     }
-    let rh = this.rowHeightCache[group.id] ? this.rowHeightCache[group.id] : 1;
-    return rh * this.props.itemHeight;
+    return this.getNonEmptyRowHeight(group.id);
+  }
+
+  getNonEmptyRowHeight(rowId) {
+    let rh = this.rowHeightCache[rowId] ? this.rowHeightCache[rowId] : 1;
+    return (
+      rh * this.props.itemHeight +
+      (rh - 1) * this.props.verticalGapBetweenOverlappingItems +
+      2 * this.props.rowTopBottomPadding
+    );
   }
 
   /**