+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [Add `scrollToItem` feature](https://github.com/flower-platform/react-timeline-10000/pull/100)
 * [Fix `onItemHover` and `onItemLeave` firing too often on items with children](https://github.com/flower-platform/react-timeline-10000/pull/98)
 * [Change color used for BackgroundGradient](https://github.com/flower-platform/react-timeline-10000/pull/96)
+* [Add `verticalGapBetweenOverlappingItems` and `rowTopBottomPadding` properties](https://github.com/flower-platform/react-timeline-10000/pull/97)
 
 ## v2.10.0